+++ tests/playwright_tests/test_select_multiple.py
@@ -5,17 +5,16 @@ from django.views.generic import FormView
 from playwright.sync_api import expect
 
 from mizdb_tomselect.views import AutocompleteView
-from mizdb_tomselect.widgets import MIZSelect
+from mizdb_tomselect.widgets import MIZSelectMultiple
 from tests.testapp.models import Person
 
 
 class MultipleForm(forms.Form):
     field = forms.ModelChoiceField(
         Person.objects.all(),
-        widget=MIZSelect(
+        widget=MIZSelectMultiple(
             model=Person,
             url="autocomplete",
-            multiple=True,
             search_lookup="full_name__icontains",
             label_field="full_name",
         ),